requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.9
django-scheduler==0.10.1
dotenv-python==0.0.1
frozenlist==1.4.1
icalendar==5.0.11
idna==3.6
importlib-metadata==7.0.1
multidict==6.0.4
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
pyee==8.2.2
pyppeteer==1.0.2
pyTelegramBotAPI==4.14.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telebot==0.0.5
telegraph-api==1.1.0
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==1.26.18
websockets==10.4
yarl==1.9.4
zipp==3.17.0
